complete file upload function

+++ Entities/Test_Result.py
@@ -7,7 +7,7 @@ class Test_Result(db.Model):
     tp_id = db.Column(db.Integer, db.ForeignKey("TestPaper.tp_id", ondelete="CASCADE"), primary_key=True)
     tp_question_no = db.Column(db.Integer, db.ForeignKey("TestPaper.tp_question_no", ondelete="CASCADE"), primary_key=True)
     s_id = db.Column(db.Integer, db.ForeignKey("Student.s_id", ondelete="CASCADE"), primary_key=True)
-    tr_mark = db.Column(db.Integer, nullable = False)
+    tr_mark = db.Column(db.Float, nullable = False)
 
     def get_cl_id(self):
         return self.cl_id
